data-science/requirements.txt
google-adk>=1.5.0
uvicorn[standard]>=0.24.0
immutabledict>=4.2.1
sqlglot>=26.10.1
db-dtypes>=1.4.2
regex>=2024.11.6
tabulate>=0.9.0
google-cloud-aiplatform[adk,agent-engines]>=1.93.0
absl-py>=2.2.2
pydantic>=2.11.3
pandas>=2.3.0
numpy>=2.3.1